compat: Move compatibility traits and types from lib.rs
This makes the public interface more clear (and easy to manage) in the
crate source code by removing the artificial private module introduced
with the previous commit.  Rename private::Rule to PrivateRule.

Use crate:: instead of super:: when importing types.
