+++ src/fs.rs
@@ -1,4 +1,4 @@
-use super::{private, uapi, Compatibility, Rule, ABI};
+use crate::{uapi, Compatibility, PrivateRule, Rule, TryCompat, ABI};
 use enumflags2::{bitflags, make_bitflags, BitFlags};
 use std::io::Error;
 use std::marker::PhantomData;
@@ -86,7 +86,7 @@ impl PathBeneath<'_> {
     }
 }
 
-impl private::TryCompat for PathBeneath<'_> {
+impl TryCompat for PathBeneath<'_> {
     fn try_compat(mut self, compat: &mut Compatibility) -> Result<Self, Error> {
         self.attr.allowed_access = self.allowed_access.try_compat(compat)?.bits();
         Ok(self)
@@ -97,7 +97,7 @@ impl private::TryCompat for PathBeneath<'_> {
 // of doing it generically.
 impl Rule for PathBeneath<'_> {}
 
-impl private::Rule for PathBeneath<'_> {
+impl PrivateRule for PathBeneath<'_> {
     fn as_ptr(&self) -> *const libc::c_void {
         &self.attr as *const _ as _
     }
